+++ litho-testing/src/main/java/com/facebook/litho/testing/logging/TestComponentsLogger.java
@@ -16,24 +16,25 @@
 
 package com.facebook.litho.testing.logging;
 
-import android.util.Pair;
-import com.facebook.litho.BaseComponentsLogger;
+import androidx.annotation.Nullable;
 import com.facebook.litho.ComponentContext;
+import com.facebook.litho.ComponentsLogger;
 import com.facebook.litho.FrameworkLogEvents;
 import com.facebook.litho.PerfEvent;
 import com.facebook.litho.TestPerfEvent;
+import com.facebook.litho.TreeProps;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 
 /**
  * In test environments we don't want to recycle the events as mockity hold on to them. We therefor
  * override log() to not call release.
  */
-public class TestComponentsLogger extends BaseComponentsLogger {
+public class TestComponentsLogger implements ComponentsLogger {
 
   private final List<PerfEvent> mLoggedPerfEvents = new LinkedList<>();
   private final List<PerfEvent> mCanceledPerfEvents = new LinkedList<>();
-  private final List<Pair<LogLevel, String>> mLoggedMessages = new LinkedList<>();
 
   @Override
   public boolean isTracing(PerfEvent logEvent) {
@@ -56,16 +57,6 @@ public class TestComponentsLogger extends BaseComponentsLogger {
     mLoggedPerfEvents.add(event);
   }
 
-  @Override
-  public void emitMessage(LogLevel level, String message) {
-    mLoggedMessages.add(new Pair<>(level, message));
-  }
-
-  @Override
-  public void emitMessage(LogLevel level, String message, int samplingFrequency) {
-    mLoggedMessages.add(new Pair<>(level, message));
-  }
-
   public List<PerfEvent> getLoggedPerfEvents() {
     return mLoggedPerfEvents;
   }
@@ -74,11 +65,13 @@ public class TestComponentsLogger extends BaseComponentsLogger {
     return mCanceledPerfEvents;
   }
 
-  public List<Pair<LogLevel, String>> getLoggedMessages() {
-    return mLoggedMessages;
-  }
-
   public void reset() {
     mLoggedPerfEvents.clear();
+  }
+
+  @Nullable
+  @Override
+  public Map<String, String> getExtraAnnotations(TreeProps treeProps) {
+    return null;
   }
 }