+++ sample-kotlin/src/main/java/com/fblitho/lithoktsample/logging/SampleComponentsLogger.kt
@@ -12,7 +12,6 @@
 package com.fblitho.lithoktsample.logging
 
 import android.util.Log
-import com.facebook.litho.BaseComponentsLogger
 import com.facebook.litho.ComponentContext
 import com.facebook.litho.ComponentsLogger
 import com.facebook.litho.FrameworkLogEvents
@@ -20,7 +19,7 @@ import com.facebook.litho.PerfEvent
 import com.facebook.litho.TreeProps
 import java.util.concurrent.atomic.AtomicInteger
 
-class SampleComponentsLogger : BaseComponentsLogger() {
+class SampleComponentsLogger : ComponentsLogger {
     private val tag = "LITHOSAMPLE"
     private val instanceCounter: AtomicInteger = AtomicInteger(0)
 
@@ -34,21 +33,6 @@ class SampleComponentsLogger : BaseComponentsLogger() {
 
     override fun cancelPerfEvent(event: PerfEvent) {
         PerfEventStore.release(event)
-    }
-
-    override fun emitMessage(level: ComponentsLogger.LogLevel, message: String) {
-        when (level) {
-            ComponentsLogger.LogLevel.WARNING -> {
-                Log.w(tag, message)
-            }
-            ComponentsLogger.LogLevel.ERROR   -> {
-                Log.e(tag, message)
-            }
-        }
-    }
-
-    override fun emitMessage(level: ComponentsLogger.LogLevel, message: String, samplingFrequency: Int) {
-        emitMessage(level, message)
     }
 
     override fun isTracing(logEvent: PerfEvent): Boolean = true
@@ -92,16 +76,16 @@ class SampleComponentsLogger : BaseComponentsLogger() {
 
     private fun getEventNameById(@FrameworkLogEvents.LogEventId markerId: Int): String =
             when (markerId) {
-                FrameworkLogEvents.EVENT_LAYOUT_CALCULATE                  -> "LAYOUT_CALCULATE"
-                FrameworkLogEvents.EVENT_MOUNT                             -> "MOUNT"
-                FrameworkLogEvents.EVENT_PRE_ALLOCATE_MOUNT_CONTENT        -> "PRE_ALLOCATE_MOUNT_CONTENT"
-                FrameworkLogEvents.EVENT_SECTIONS_CREATE_NEW_TREE          -> "SECTIONS_CREATE_NEW_TREE"
+                FrameworkLogEvents.EVENT_LAYOUT_CALCULATE -> "LAYOUT_CALCULATE"
+                FrameworkLogEvents.EVENT_MOUNT -> "MOUNT"
+                FrameworkLogEvents.EVENT_PRE_ALLOCATE_MOUNT_CONTENT -> "PRE_ALLOCATE_MOUNT_CONTENT"
+                FrameworkLogEvents.EVENT_SECTIONS_CREATE_NEW_TREE -> "SECTIONS_CREATE_NEW_TREE"
                 FrameworkLogEvents.EVENT_SECTIONS_DATA_DIFF_CALCULATE_DIFF -> "SECTIONS_DATA_DIFF_CALCULATE_DIFF"
-                FrameworkLogEvents.EVENT_SECTIONS_GENERATE_CHANGESET       -> "SECTIONS_GENERATE_CHANGESET"
-                FrameworkLogEvents.EVENT_SECTIONS_ON_CREATE_CHILDREN       -> "SECTIONS_ON_CREATE_CHILDREN"
-                FrameworkLogEvents.EVENT_SECTIONS_SET_ROOT                 -> "SECTIONS_SET_ROOT"
-                FrameworkLogEvents.EVENT_CALCULATE_LAYOUT_STATE            -> "CALCULATE_LAYOUT_STATE"
-                else                                                       -> "UNKNOWN"
+                FrameworkLogEvents.EVENT_SECTIONS_GENERATE_CHANGESET -> "SECTIONS_GENERATE_CHANGESET"
+                FrameworkLogEvents.EVENT_SECTIONS_ON_CREATE_CHILDREN -> "SECTIONS_ON_CREATE_CHILDREN"
+                FrameworkLogEvents.EVENT_SECTIONS_SET_ROOT -> "SECTIONS_SET_ROOT"
+                FrameworkLogEvents.EVENT_CALCULATE_LAYOUT_STATE -> "CALCULATE_LAYOUT_STATE"
+                else -> "UNKNOWN"
             }
 
     private fun Float.withDigits(digits: Int) = java.lang.String.format("%.${digits}f", this)